backend/requirements.txt
# Django Core
Django==5.0.1
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# CORS Headers
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# Image Processing
Pillow==10.2.0

# Cloudinary
cloudinary==1.38.0
django-cloudinary-storage==0.3.0

# Email
sendgrid==6.11.0

# Security
django-ratelimit==4.1.0
bleach==6.1.0

# Server
gunicorn==21.2.0

# Code Quality
black==24.1.1
ruff==0.1.14

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# API Documentation
drf-spectacular==0.27.0
